Format document
VsCode and other tools can format your js to be uniform. These settings can be configured.
Note: This is going to make it look like a lot of stuff changed, but in reality only the whitespace and stuff did. It is better to do this early on so that future formatted changes don't change lines that weren't touched.
Per the original formatting, we will be using 4 spaces as tab.