Modify playRound to return result instead of message.
playRound's purpose is actually to determine the winner, but it does not need to decide what to do about it. By having the response be a uniform win/lose/tie, we can decide how to deal with it however we want. Also, if languages were to change or we wanted different response messages, then this allows us to do so.